Organized websocket code
